Create contact_rollups_raw model and move extraction logic there
